Slide to a standardized spot
The computed chunk is now moved to a standard spot so that I can treat them more canonically. One corner is sent to the origin.

There was a somewhat major mistake in how I was calculating chosen_origin, the corner which gets put on the origin, but I wasn't using it for anything before apparently, so I didn't notice.

There was also an extra point , [0,0,0,0,0,0], which I was always leaving in the list of chunks --- surprised that didn't cause any issues, too.