Actually got randomized version working OK
Currently there are some "bad runs" with errors I haven't gotten to the bottom of, but I handle everything somewhat gracefully and the results from the "good runs" seem useable. Only issue is that it takes a long time. Did a run getting 4,000 constraints and saw no sign that I was filling up the constraint space.