Search os blocks, print more info
Seems as if part of the problem was the templates I'm starting with already having certain blocks in os_blocks which need to be in is_blocks, so, I now look in both places.

Mysteriously, I've gotten rid of the "duplicate chunk" and "outside hit" errors, yet still have trouble locating the player within the grid.