New formula for chunk containment
My "safely_contains" function, which is called so frequently in searching for the player, had a mistake that's obvious in hindsight.

There are a few glaring logical problems in the algorithm now, but it seems to work okay (although slow). Main todos: recognize blocks' diagonal neighbors, and write a valid check for whether some ensemble of 'might-contain' chunks is guaranteed to have the player's location covered.